+++ src/main/java/com/thread/concurrency/counter/queue/Producer.java
@@ -1,4 +1,4 @@
-package com.thread.concurrency.counter.queueCounter;
+package com.thread.concurrency.counter.queue;
 
 public interface Producer {
     void add(long value) throws InterruptedException;
